scraper/requirements.txt
praw==7.8.1
python-dotenv==1.0.0
fluent-logger==0.11.1